Set OG images for pages
