+++ website/docs/components/models/index.md
@@ -2,6 +2,7 @@
 title: 'Model Providers'
 sidebar_label: 'Model Providers'
 description: 'Overview of supported model providers for ML and LLMs in Spice.'
+image: /img/og/models.png
 ---
 
 Spice supports various model providers for traditional machine learning (ML) models and large language models (LLMs).
