+++ website/src/pages/cookbook/index.mdx
@@ -1,0 +1,10 @@
+---
+title: Spice.ai OSS Cookbook
+descrtiption: A collection of guides and samples to help you build data-grounded AI apps and agents with Spice.ai Open-Source. Find ready-to-use examples for data acceleration, AI agents, LLM memory, and more.
+image: /img/og/cookbook.png
+hide_table_of_contents: true
+---
+
+import {CookbookPage} from './cookbook-page'
+
+<CookbookPage />
